requirements.txt
# Locopon - Swedish Deals Intelligence System
# Core Dependencies

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# HTTP Client
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# AI Integration
openai>=1.3.0

# Telegram Bot
python-telegram-bot>=20.6

# Environment Management  
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.4

# Logging
loguru>=0.7.2

# JSON Processing
orjson>=3.9.10

# Date/Time
python-dateutil>=2.8.2

# Validation
pydantic>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Development
black>=23.9.1
flake8>=6.0.0
mypy>=1.6.0
